Factor out empty lesson plan message
